function updateHighestBidDetails to not only update highest bid in bidhistory but also in the details column. Also made 'start your selling journey' functional.
