+++ index.html
@@ -389,9 +389,12 @@
               Look no further. Provide us with information and images, and we
               will take care of the rest.
             </p>
-            <button class="btn btn-primary mt-3">
+            <a
+              href="src/html/sell-page/index.html"
+              class="btn btn-primary mt-3"
+            >
               Start your selling journey
-            </button>
+            </a>
           </div>
           <div class="col-md-6">
             <img
