+++ src/html/sell-page/index.html
@@ -81,12 +81,15 @@
         </nav>
       </header>
       <main>
-        <div class="container py-5 h-100" id="createListingContainer"></div>
-        <div
-          id="successMessage"
-          class="alert alert-success"
-          style="display: none"
-        ></div>
+        <div class="container py-3">
+          <h1>Create a listing</h1>
+          <div class="container" id="createListingContainer"></div>
+          <div
+            id="successMessage"
+            class="alert alert-success"
+            style="display: none"
+          ></div>
+        </div>
       </main>
       <div class="container my-5">
         <footer class="bg-light text-center text-lg-start text-dark">
